@@ -1,26 +1,25 @@
-import React from 'react';
+import logo from './logo.svg';
 import './App.css';
-import {BrowserRouter, Routes, Route} from 'react-router-dom';
-import Login from './pages/login';
-import SignUp from './pages/sign-up';
-import Home from './pages/home';
-import Model3DDashboard from './pages/my-model3ds';
-import ErrorPage from './pages/error-page';
 
-const App = () => {
+function App() {
   return (
-    <div className="App bg-black">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Home />} />
-          <Route exact path="/my-model3ds" element={<Model3DDashboard />} />
-          <Route exact path="/login" element={<Login />} />
-          <Route exact path="/sign-up" element={<SignUp />} />
-          <Route exact={true} path="*" element={<ErrorPage />} />
-        </Routes>
-      </BrowserRouter>
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.js</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
     </div>
   );
-};
+}
 
 export default App;
